Moved dflowfm test files into a newly created dflowfm subfolder.
