+++ tests/io/dflowfm/test_polyfile.py
@@ -22,7 +22,7 @@ from hydrolib.core.io.dflowfm.polyfile.parser import (
 )
 from hydrolib.core.io.dflowfm.polyfile.serializer import Serializer, write_polyfile
 
-from ..utils import assert_files_equal, test_input_dir, test_output_dir
+from ...utils import assert_files_equal, test_input_dir, test_output_dir
 
 
 class TestSerializer:
