+++ tests/io/dflowfm/test_structure.py
@@ -27,7 +27,7 @@ from hydrolib.core.io.dflowfm.structure.models import (
     Weir,
 )
 
-from ..utils import WrapperTest, invalid_test_data_dir, test_data_dir
+from ...utils import WrapperTest, invalid_test_data_dir, test_data_dir
 
 uniqueid_str = "Unique structure id (max. 256 characters)."
